authselect: add custom/local policy
This is a copy of the 'minimal' policy. The implementation
needs some work. I'm not sure how to manage a custom policy
cleanly and separate the 'local' code (ie. the custom policy)
from the base authselect module. Also, the custom policy is
optional so we need to handle that correctly.